removing now the old `Simulator`. this requires a change in update
module because this contains now the `ScriptApi`. I hope we can get
the structure repaired again -> this should be in the app-module;
also, I hope we get rid of the script based updates soon.

the nice thing is, that we do not need to run a first calculation
to initialize the simulation:

```java
Simulator solver = new Simulator(setup, matrixCache, App.getSolver());
// do a first calculation that initialises the result;
solver.nextRun();
```

also, let me remove some trivial classes ... we can just call dispose
on the progress section -> results in a much better layout