the pinned contributions should be stored in the `Simulator` and
the `SimulationResult`. Let me add a `PinBoard`. Later, we can
make the font bold for pinned providers. Also, when sorting them
we should always have the pinned providers at the top and they
should never disappear when filtering the list.

The simulator should have the methods:

* `getFlowIndex(): FlowIndex`
* `getTechIndex(): TechIndex`

and we use them in the editor instead going to the result.

also, it needs a public list `pinnedProviders` (or similar)