here are some tests with the new Rust based native library
bindings in the `systems2` branch: we run a Monte-Carlo
simulation with a parameterized ecoinvent 3 system and
LCIA calculation. Here are the results:

`systems2` without UMFPACK bindings (plain dense with
OpenBLAS):

* time ~ 42 seconds
* RAM usage ~ 4.8 GB

`systems2` with UMFPACK bindings:

* time ~ 27 seconds
* RAM usage ~ 1.2 GB

note that the matrix with the LCIA factors is probably dense.

In comparison with the current 1.8.0 release:

* time ~ 46 seconds (+ some seconds as the first calculation
  is done on initialization)
* RAM usage ~ 5.5 GB