fixed #74
It was even a bit more complicated than I first thought. If it is a
global
parameter it is fine to just update the `version` and `lastChange`
fields.
However, when it is a parameter in a process or LCIA method we also need
to update the entity to which the parameter belongs.

Also, when a global parameter or the entity to which a parameter belongs
is currently edited in another editor it should not be possible to
modify
that parameter from the parameter overview.