for the macOS build this structure works
with the Eclipse executable
(from the official Eclipse RCP package):

```
openLCA.app
 - Contents
   - Eclipse
     - configuration
     - plugins
     - .eclipseproduct
     - eclipse.ini
   - MacOS
     - eclipse
   - Resources
   - Info.plist
```

In the `Info.plist` everything needs to be renamed
to `Eclipse` except the executable is `eclipse`.
Also, in the ini file the startup and launcher
paths need to be added:

```
-startup
../Eclipse/plugins/org.eclipse.equinox.launcher...jar
--launcher.library
../Eclipse/plugins/org.eclipse.equinox...
```

just moving around the folders does not work as
the ini file is not picked up. Now, try again with the
build output ... it seems to work only with the structure
above