a decision table / tree for finding new link candidates
in the product system editor;

https://en.wikipedia.org/wiki/Decision_table
("Entscheidungstabelle")

For a given exchange `e` of a process or sub-system
in a product system we want to check for each possible
link candidate `c` which describes another exchange if
we can connect these both exchanges.

We assume that each candidate `c` has the opposite
direction than `e` (so when `e` is an input then `c` is an
output) and that both exchanges have the same flow. (Also,
we assume that the processes or sub-systems of `e` and `c`
are different.)?

Whether we can connect both exchanges depends on the
following conditions:

* `e.Input`: whether `e` is an input or not
* `Waste`: whether the flow is a waste flow
* `e.Linked`: whether `e` is already linked
* `c.Linked`: whether `c` is already linked

This results in `2^4 = 16` combinations but not
every condition is relevant in each case:

```
e.Input  Waste  e.Linked  c.Linked  |  CanConnect
   0       0        _         0     |       1
   0       0        _         1     |       0
   0       1        0         _     |       1
   0       1        1         _     |       0
   1       0        0         _     |       1
   1       0        1         _     |       0
   1       1        _         0     |       1
   1       1        _         1     |       0
```

We can further reduce the table by combining
the conditions `e.Input` and `Waste` to a
condition `e.Provider` which are product outputs
or waste inputs:

```
e.Provider = e.Input == Waste
```

```
 e.Provider  e.Linked  c.Linked  |  CanConnect
      1          _         0     |       1
      1          _         1     |       0
      0          1         _     |       0
      0          0         _     |       1
```

With this:

```
CanConnect = (e.Provider && !c.Linked)
              || (!e.Provider && !e.Linked)
```