thinking about it: the new flow mapping tool can fully replace the
current bulk-replace tool as the user can create new flow mappings in
the tool and directly apply them.

A menu contribution of the mapping tool would be good with the
following entries:

* Open from JSON-LD package
* Initialize from JSON-LD package
* Initialize from ILCD package
* Save as JSON-LD package
* Save as CSV file

The `Initialize ...` functions will initialize/propose a mapping for
the flows in data package based on flow descriptors like name, category,
description etc.

To also handle product flows, the mapping editor should include the
mapping of `default providers` for product flows... This will result
in a table with quite some columns. For editing of the mapping entries
a dialog would be good so that not all information (like UUIDs,
flow properties etc.) need to be shown in the table.

The replace logic needs to consider very different cases:

* replace existing flows with flows that are not yet in the database ->
  we need implementations of something like an IFlowProvider interface
  for the database, JSON-LD packages, and ILCD packages
* the flows may be replaced in different locations: processes and|or
  LCIA methods
* fully replaced flows may be deleted after the mapping (if they are
  not used anymore)
* maybe the user wants to only map the default providers, and the new
  default providers may not yet exist in the database: again, we would
  need something like implementations of an IProcessProvider interface
* and more ...