the JsonProvider should construct the flow map.
add the following things:

* category paths
* flow types (also in fedmapld)
* unit + flow property (also in fedmapld)
* (locations)

then write a first version of the FlowReplacer
with updateable cursors ... there are many different
cases to consider when we also include product flows,
e.g. quantitative references of product systems
& project variants

...

for now we will assume that the unit and flow property
is set in the mapping.