for generating a flow mapping we first need to
collect all flows from the ILCD package. When
no CSV mapping is provided, we generate a mapping
based on the flows in the database:

for a flow `s` in the ILCD package we want to
find a matching flow `t` in the database. We
compare the following attributes:

* UUID => dominates all other attributes
* Name => a syntax aware matching function
* Category => stemming and longest common path
* Flow property & unit

we could start with something like a scoring
system:

```
for (FlowRef s : provider.getFlowRefs()) {
  FlowRef candidate = null;
  double score = 0.0;
  for (FlowRef t : db.getFlowRefs()) {
    if (Match.sameUUID(s, t)) {
      candidate = t;
      break;
    }
    double ns = Match.nameScore(s, t);
    double cs = Match.categoryScore(s, t);
    double us = Match.unitScore(s, t);
    doube s = 100 * ns + 10 * cs + us;
    if (candidate == null || s > score) {
      candidate = t;
      score = s;
    }
  }
  if (candidate != null) {
    // add to mapping
  }
}
```

we need also some functionality in the table to
find and select candidates with a score from
the database where the user can select another
flow...

but first we need to implement and test the
matcher ...

sketching the general workflow:

```
=> open ILCD package
  => generate flow mapping with flows in db
    => open this in the editor.
```

```
=> MappingTool.openFile()
  => MappingTool.generate()
    => MappingTool.open()
```

... just learned that source and target is not so
clear currently:

* when we have a JSON-LD package we want to replace
  the flows in the database with the flows in the
  JSON-LD package: source is the database and target
  the JSON-LD package
* when we have an ILCD package we want to map the
  flows in the ILCD package to the flows in the
  databse: source is the ILCD package and target the
  database

maybe we could implement a `swap` function that
switches source and target flows in a flow map?