redesign the ILCD and JSON-LD providers
On `sync` and `persist` the provider should open and close the
underlying zip-file => with this we do not need to close the
provider when the editor is closed (and leak resources when
we forget this).

```java
try (store = ZipStore.open(file)) {
  // do something
} catch (Exception e) {
  // handle errors
}
```

the sync functions are currently really inconsistent... but there
are so many places where errors can occur and we have to define
what we catch/check where and present to the user.