now in the import we can use custom mapping files
we should update (later) the SimaPro CSV import so that it
uses the same format. Also, we may want to allow file headers
and switch to commas `,` as column separators instead of
semicolons.

regarding the descriptors: we should throw out the
`CategorizedDescriptor` type... it is not always helpful to
put everything into a class hierarchy: a thing can have a
category or not and then the field is just `null`.

next steps:

* simplify the descriptors
* build the sync functions FlowRef <-> Provider
* graphical editing of flow mappings
* functions for building mappings automatically based on
  flow attributes

a good test regarding the descriptor performance are all
the (300000!) flows from the Fed.LCA flow list...