in the editing dialog we need to work on a clone of the respective
mapping entry. the provider of the target flow can be set when
the target system is the currently active database and the
target flow is a product or waste flow. the question is then
how to set the provider in the import (first, we will only
suport this for the ILCD import).