not sure why the inner UI job (the callback) still
runs in the outer progress? -> we schedule it as
a new job, why is the progress bar still active?

now it gets even more complicated, we want to
return the selected FlowRef from the inner UI job.

the easiest thing is to just pass in a callback
and we make it explicit that maybe nothing was
selected at all.

do we get the selection from a connected REPL?

first this:

```clojure
(defn callback [optional]
  (if (.isPresent optional)
    (-> optional
      (.get)
      (.flow)
      (.name)
      (println))
    (println "nothing selected")))
```

-> `user$callback cannot be cast to java.util.function.Consumer`

but with `reify`:

```clojure
(import
  (java.io File)
  (org.openlca.app.tools.mapping.model ILCDProvider)
  (org.openlca.app.tools.mapping FlowRefDialog)
  (org.eclipse.swt.widgets Display))

(def provider
 (ILCDProvider/of
  (File. "C:/Users/ms/Downloads/EUFRP.zip")))

(def callback
  (reify java.util.function.Consumer
    (accept [this optional]
      (if (.isPresent optional)
       (-> optional
         (.get)
         (.flow)
         (.name)
         (println))
       (println "nothing selected")))))

(.asyncExec
 (Display/getDefault)
 (fn []
   (FlowRefDialog/open
     provider
     callback)))
```