hope this makes the UI of the mapping tool
a bit more clear ...

Apply on database ...

if an openLCA database is not selected as source system:

error: Source system should be a database
"""
This only works when the source system is a database where the flows
should
be replaced with flows from the target system (which can be the same
database).
"""

otherwise, who

Replace flows in database
"""
This will replace the flows in the database (the source system) with the
flows in the target system.

[ ] Replace in processes
[ ] Replace in LCIA methods
[ ] Delete replaced flows
"""