the navigation update needs to run in the UI thread
regarding the status: we could save the status for the source
and target flow into the CSV file but when opening a mapping file
the tool should indicate that no check was performed yet (e.g.
via a warning symbol on the `Check mappings` button)

now checking the ckeck function:

=> this was just a label problem!

providers are currently not checked; and now it comes: we also
need to set the providers in the replacer!