disable EcoSpold 2 import and export wizards
there is only one source that distributes EcoSpold 2 files: ecoinvent.
however, the ecoinvent data sets are not really usable in openLCA
without pre-processing (e.g. remove formulas that are artifacts from
the raw data sets but are not valid anymore in the data sets of a
linked system model, remove 0 exchange values, swap waste flows, map
elementary flows). We distribute prepared ecoinvent packages via openLCA
Nexus.

When using EcoSpold 2 as an exchange format, we should not apply these
rules (e.g. remove formulas) because we do not know if this is an
ecoinvent data set. But then, the ecoinvent data sets do not work
correctly. Thus, we remove the standard EcoSpold 2 import and export
now as it leads to more confusion than helps.