we have to decide if we support formulas in parameters
of uncertainty distributions or not. when we do this, then
a formula could contain a parameter that again has an
uncertainty distribution assigned. thinks like that could
result in strange edge cases... currently, it is partially
supported in the calculation but not in the editors... so
we could remove that feature savely.