We use 3 parameters to control the depth of the exported tree
* the maximum number of levels of the tree (the maximum length of a
path):
  $d_{max}$
* minimum contribution: $c_{min}$
* the recursion depth (max. number of repetitions of a process in a
path, this
  the max. number of loop expansions along a path): $d_{rec}$

In the user interface, $d_{max}$ and $c_{min}$ can be set to
_unlimited_.
Internally we set $d_{max}$ and $c_{min}$ to a value $< 0$ in this case.
If
$d_{max} < 0$ a value of $d_{rec} \ge 0$ is required to control the
expansion
of cycles along a path if the underlying product system has loops (in
most
cases a value $c_{min} > 0$ will also control such expansions but there
are
cases where the contribution grows along a expanded cycle __TODO:__ show
that).

Suppose we have a node $node_p$ of a process $p$ in a path $path$. We
cut the
expansion of $path$ at $node_p$ in the following cases, where
$c_{node_p}$ is
the upstream contribution of $node_p$:

* $c_{node_p} = 0$
* $d_{max} \ge 0 \ \textbf{and} \ length(path) \ge d_{max}$
* $c_{min} \ge 0 \ \textbf{and} \ abs(c_{node_p}) < c_{min}$
* $\{p^{count}: p \in path\}, count \ge d_{rec}$

Thus in order to control the depth of the tree with these parameters we
always
need the complete path to a node. We traverse the upstream tree in depth
first
order in an algorithm like this:

* $traverse(node_p, path)$
  * $\textbf{if} \ \textbf{not} \ expand(node_p, path) \
\textbf{return}$
  * $row := row + 1$
  * $column := length(path)$
  * $write(node_p, row, column)$
  * $\textbf{for} \ node_q \ \textbf{in} \ childs(tree, node_p)$
    * $traverse(node_q, append(path, node_p))$

The $expand$ function checks the conditions above. Each node is written
in a
separate row and the column of that node corresponds to the length of
the path.
To reduce memory usage, we share common prefixes between paths. A path
is just
an immutable data structure with a node and a pointer to a prefix:

```java
class Path {
  final Path prefix;
  final UpstreamNode node;

  Path append(UpstreamNode node) {
    // ...
  }
}
```

As abstract data type:

$
\textbf{datatype} \ Path: \\
\ \ new(Node) \rightarrow Path \\
\ \ append(Path, Node) \rightarrow Path \\
\ \ length(Path) \rightarrow \textbf{int} \\
\ \ count(Path, Process) \rightarrow \textbf{int}
$