removed the parameter eval. action in the dev. mode
this can be now (better) done in the "big-parameter-table". also, removed
the interpreter+scope from the uncertainty dialog. the interpreter was
always `null` so it was never possible to evaluate formulas in this dialog.
this means we could remove the formula fields from the uncertainty type as
it was never possible to set these fields in the UI and they were probably
never used elsewhere...