currently, we just cannot combine the fast/direct
process calcuation with libraries. so we will just disable it
when a library is linked to a database. next thing is the
meta-data export. the reference amounts that are stored in
the database are wrong; it should be 1 reference unit of the
reference flow. this is displayed correctly, as we fetch the
data from the library but the value in the database may
is related to another unit:

```sql
-- displays 1 MJ, but stored is 1 kWh
select u.name from tbl_exchanges e
  inner join tbl_units u on
  e.f_unit = u.id
  where f_owner = 1578548
```