we have two `viewers` packages
merged it. time to cleanup the bundle cache...

so that IntelliJ shows the correct compiler errors.
there is this issue: https://www.eclipse.org/forums/index.php/t/167161/
not sure if there is a cleanup function for the target platform...
just delete the cache and we should get the latest versions of
our plugins... this should also speed up the tools a bit.

indexing ...

now it cannot resolve the things anymore. ahh Eclipse is still
downloading stuff

now it works.