adding a flow is more complicated
first, we also want to add elementary flows here. means that we need a
switch
to display also elementary flows in the graph (with icons, and more
options
like amounts and units). but this maybe later...

first, we need to simply add flows, then add them to the selected
process,
then update the process in the graph...