try to repaint it recursively
```
var queue = new ArrayDeque<IFigure>();
queue.add(editor.getModel().figure);
while(!queue.isEmpty()) {
  var figure = queue.pop();
  figure.repaint();
  queue.addAll(figure.getChildren());
}
```

this does not work, but this:

```java
editor.getGraphicalViewer()
  .getEditPartRegistry()
  .forEach((key, part) -> {
    if (part instanceof EditPart) {
      ((EditPart) part).refresh();
    }
  });
```

how does the mass expansion does this?

```
viewer.setContents(model);
```

so it recreates the complete thing... but this is
not something we can do here.

>  Collapse all process nodes in the editor. Only the
>  reference process will be added.

but what if we do not have a reference process... in
future versions this might be the case...