bug fix: newly created products cannot be linked
ok, this is in the `CreateLinkCommand`:

```java
FlowType type = sysNode.flows.type(flowId);
```

the system node holds the flows, and they are not updated when a new
flow is added to the system...

also, add the connection layer on top of the model layer. as the root
node is now selectable it is otherwise not possible to select a
connection. this results in node crossing but which is not so a problem
when the routing property is selected, which should anyhow the default.