use plain `JsonObject` in `GraphFile`
because we want to save other settings in that same file. the streaming
approach was fast but it relied on the same order of fields. adding
a settings object to the top would break the reader. so switching to
plain `JsonObject` parsing for now. we can switch back to some streaming
(but then with field order checking) if the performance of this is too
slow...