#125 the only way to get the trees faster is
to reduce the number of root elements. when there are more than 50 items
on the root level, we could try to categorize them. this should make the
tree much faster. searching would be also cool. this is getting complex,
so moving this into a separate package with specific item types.