for now we just do not allow to
add a library to a database where the library processes are already
contained. later, however, we want to support that a mounted library
can be replaced by another library that has the same interface (e.g.
by an updated version but also different allocation models, regionalized
versions etc.). in this case, just the _used_ interface has to be the same...