do not show the direct calculation button in
the toolbar when the database is linked to a library. this does not
work. but
what is if we want to run gRPC calculations on a single created process?
we
need to create that product system then and link it to the library
processes.
then, do we delete it after the calculation? can we create it in memory
only?
that would be better.