it does not select the reference process in the
wizard when you create a product system directly from a library process.
it
finds the element in the navigation:

```java
var descriptor = Descriptor.of(refProcess);
var elem = Navigator.findElement(descriptor);
```

but in `checkInput` the reference process is null:

```java
if (refProcess == null) {
  ...
```

seems that `processTree.setSelection(new StructuredSelection(elem));`
fails.

maybe this is the problem: we have two instances of `ModelTypeElement`
with
the same model type in the tree: one at the root level and one or more
under
the libraries. when we compare navaigation elements we cannot just
compare the
content, we need to compare the path recursively.