regio-lcia: move the collection of attributes
from features to the `GeoParam` class. this is a bit clearer. also, we
want to
create a fresh setup from a GeoJSON file or restore it from a JSON file
in the
`Setup.read` method... we simply detect a setup file when the JSON
object has
a `setup` and `features` attribute, otherwise we parse it a GeoJSON
file.