sort databases names with spaces correctly
we had this before:

#+begin_src java
// when a database is connected to a repository, the navigation label
// contains the repository URL as suffix; we do not want to include
// that URL in this comparison
if (e1 instanceof DatabaseElement && name1.contains(" ")) {
    name1 = name1.substring(0, name1.indexOf(" "));
}
if (e2 instanceof DatabaseElement && name2.contains(" ")) {
    name2 = name2.substring(0, name2.indexOf(" "));
}
#+end_src

but we do not really need this. The database names are different
be definitions so the URL suffix does not matter. Also, if they
were the same, why not sorting them by the URL suffix.