Update soda-nodes.json
Adding "EF Representative Products" node and "EF Small data provider" node. Changing name of "Small Data Providers Database" to "ILCD-EL Small Data Providers Database" so that it is clear that this is the old EF node with ILCD-EL compliance and there is a newer EU node for small data providers with the EF compliance data.